dweeter/requirements.txt
Django==1.8
MySQL-python==1.2.3
djangorestframework